Updated ranger and enemybullet to correctly interact with healthbehavior. also updated GUIBar to show a percantage, rather than a max amount

+++ Assets/GUI/GUIBar.cs
@@ -21,7 +21,7 @@ public class GUIBar : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        barDisplay = player.getCurrentHP()/player.getMaxHP();//divided by max HP to get num between 0 and 1
+        barDisplay = (float)player.getCurrentHP()/player.getMaxHP();//divided by max HP to get num between 0 and 1
         
     }
